refactoring every tests for the module alignment methods in a common file
